refactor: Create WordPress parser
